Moving tf/core/blocks to tf/core (#565)
