+++ merlin/models/tf/blocks/cross.py
@@ -17,8 +17,8 @@ from typing import List, Optional, Tuple, Union
 
 import tensorflow as tf
 
-from merlin.models.tf.blocks.core.combinators import Filter, SequentialBlock, TabularBlock
 from merlin.models.tf.blocks.mlp import DenseMaybeLowRank, InitializerType, RegularizerType
+from merlin.models.tf.core.combinators import Filter, SequentialBlock, TabularBlock
 from merlin.models.tf.utils.tf_utils import (
     maybe_deserialize_keras_objects,
     maybe_serialize_keras_objects,
