+++ merlin/models/tf/blocks/dlrm.py
@@ -16,9 +16,9 @@
 
 from typing import Optional
 
-from merlin.models.tf.blocks.core.base import Block, Debug
-from merlin.models.tf.blocks.core.combinators import Filter, ParallelBlock, SequentialBlock
 from merlin.models.tf.blocks.interaction import DotProductInteraction
+from merlin.models.tf.core.base import Block, Debug
+from merlin.models.tf.core.combinators import Filter, ParallelBlock, SequentialBlock
 from merlin.models.tf.inputs.continuous import ContinuousFeatures
 from merlin.models.tf.inputs.embedding import EmbeddingFeatures, EmbeddingOptions
 from merlin.schema import Schema, Tags
