+++ merlin/models/tf/blocks/mlp.py
@@ -18,9 +18,9 @@ from typing import List, Optional, Union
 
 import tensorflow as tf
 
-from merlin.models.tf.blocks.core.base import Block
-from merlin.models.tf.blocks.core.combinators import ResidualBlock, SequentialBlock
-from merlin.models.tf.blocks.core.tabular import Filter, tabular_aggregation_registry
+from merlin.models.tf.core.base import Block
+from merlin.models.tf.core.combinators import ResidualBlock, SequentialBlock
+from merlin.models.tf.core.tabular import Filter, tabular_aggregation_registry
 from merlin.models.tf.utils.tf_utils import (
     maybe_deserialize_keras_objects,
     maybe_serialize_keras_objects,
