+++ merlin/models/tf/blocks/retrieval/matrix_factorization.py
@@ -18,9 +18,9 @@ from typing import Any, Callable, Dict, Optional, Union
 
 import tensorflow as tf
 
-from merlin.models.tf.blocks.core.aggregation import CosineSimilarity
-from merlin.models.tf.blocks.core.transformations import RenameFeatures
 from merlin.models.tf.blocks.retrieval.base import DualEncoderBlock
+from merlin.models.tf.core.aggregation import CosineSimilarity
+from merlin.models.tf.core.transformations import RenameFeatures
 from merlin.models.tf.inputs.embedding import EmbeddingFeatures, EmbeddingOptions
 from merlin.schema import Schema, Tags, TagsType
