+++ merlin/models/tf/core/combinators.py
@@ -7,7 +7,7 @@ import six
 import tensorflow as tf
 from tensorflow.keras.layers import Layer
 
-from merlin.models.tf.blocks.core.base import (
+from merlin.models.tf.core.base import (
     Block,
     BlockType,
     NoOp,
@@ -15,7 +15,7 @@ from merlin.models.tf.blocks.core.base import (
     is_input_block,
     right_shift_layer,
 )
-from merlin.models.tf.blocks.core.tabular import Filter, TabularAggregationType, TabularBlock
+from merlin.models.tf.core.tabular import Filter, TabularAggregationType, TabularBlock
 from merlin.models.tf.utils import tf_utils
 from merlin.models.tf.utils.tf_utils import call_layer
 from merlin.models.utils import schema_utils
@@ -592,7 +592,7 @@ class ResidualBlock(WithShortcut):
         strict: bool = False,
         **kwargs,
     ):
-        from merlin.models.tf.blocks.core.aggregation import SumResidual
+        from merlin.models.tf.core.aggregation import SumResidual
 
         super().__init__(
             block,
