+++ merlin/models/tf/core/prediction.py
