+++ merlin/models/tf/inputs/base.py
@@ -17,10 +17,10 @@
 import logging
 from typing import Dict, Optional, Tuple, Type, Union
 
-from merlin.models.tf.blocks.core.aggregation import SequenceAggregation, SequenceAggregator
-from merlin.models.tf.blocks.core.base import Block, BlockType
-from merlin.models.tf.blocks.core.combinators import ParallelBlock, TabularAggregationType
-from merlin.models.tf.blocks.core.transformations import AsDenseFeatures
+from merlin.models.tf.core.aggregation import SequenceAggregation, SequenceAggregator
+from merlin.models.tf.core.base import Block, BlockType
+from merlin.models.tf.core.combinators import ParallelBlock, TabularAggregationType
+from merlin.models.tf.core.transformations import AsDenseFeatures
 from merlin.models.tf.inputs.continuous import ContinuousFeatures
 from merlin.models.tf.inputs.embedding import (
     ContinuousEmbedding,
