+++ merlin/models/tf/models/base.py
@@ -11,11 +11,11 @@ from keras.utils.losses_utils import cast_losses_to_common_dtype
 from tensorflow.keras.utils import unpack_x_y_sample_weight
 
 import merlin.io
-from merlin.models.tf.blocks.core.base import Block, ModelContext, PredictionOutput, is_input_block
-from merlin.models.tf.blocks.core.combinators import SequentialBlock
-from merlin.models.tf.blocks.core.prediction import Prediction, PredictionContext
-from merlin.models.tf.blocks.core.tabular import TabularBlock
-from merlin.models.tf.blocks.core.transformations import AsDenseFeatures, AsRaggedFeatures
+from merlin.models.tf.core.base import Block, ModelContext, PredictionOutput, is_input_block
+from merlin.models.tf.core.combinators import SequentialBlock
+from merlin.models.tf.core.prediction import Prediction, PredictionContext
+from merlin.models.tf.core.tabular import TabularBlock
+from merlin.models.tf.core.transformations import AsDenseFeatures, AsRaggedFeatures
 from merlin.models.tf.dataset import BatchedDataset
 from merlin.models.tf.inputs.base import InputBlock
 from merlin.models.tf.losses.base import loss_registry
@@ -28,7 +28,7 @@ from merlin.models.utils.dataset import unique_rows_by_features
 from merlin.schema import Schema, Tags
 
 if TYPE_CHECKING:
-    from merlin.models.tf.blocks.core.index import TopKIndexBlock
+    from merlin.models.tf.core.index import TopKIndexBlock
 
 
 class MetricsComputeCallback(tf.keras.callbacks.Callback):
@@ -854,7 +854,7 @@ class RetrievalModel(Model):
         self.has_item_corpus = False
 
         if item_corpus:
-            from merlin.models.tf.blocks.core.index import TopKIndexBlock
+            from merlin.models.tf.core.index import TopKIndexBlock
 
             self.has_item_corpus = True
