+++ merlin/models/tf/models/benchmark.py
@@ -15,13 +15,13 @@
 #
 from typing import List, Optional, Union
 
-from merlin.models.tf.blocks.core.aggregation import ElementWiseMultiply
-from merlin.models.tf.blocks.core.base import Block
-from merlin.models.tf.blocks.core.combinators import ParallelBlock
 from merlin.models.tf.blocks.retrieval.matrix_factorization import (
     MatrixFactorizationBlock,
     QueryItemIdsEmbeddingsBlock,
 )
+from merlin.models.tf.core.aggregation import ElementWiseMultiply
+from merlin.models.tf.core.base import Block
+from merlin.models.tf.core.combinators import ParallelBlock
 from merlin.models.tf.models.base import Model
 from merlin.models.tf.models.utils import parse_prediction_tasks
 from merlin.models.tf.prediction_tasks.base import ParallelPredictionBlock, PredictionTask
