+++ merlin/models/tf/prediction_tasks/base.py
@@ -4,7 +4,7 @@ import tensorflow as tf
 from keras.utils.generic_utils import to_snake_case
 from tensorflow.keras.layers import Layer
 
-from merlin.models.tf.blocks.core.base import (
+from merlin.models.tf.core.base import (
     Block,
     BlockType,
     ContextMixin,
@@ -12,7 +12,7 @@ from merlin.models.tf.blocks.core.base import (
     PredictionOutput,
     name_fn,
 )
-from merlin.models.tf.blocks.core.combinators import ParallelBlock
+from merlin.models.tf.core.combinators import ParallelBlock
 from merlin.models.tf.typing import TabularData, TensorOrTabularData
 from merlin.models.tf.utils import tf_utils
 from merlin.models.utils.schema_utils import tensorflow_metadata_json_to_schema
