+++ merlin/models/tf/prediction_tasks/next_item.py
@@ -19,16 +19,16 @@ from typing import Optional
 import tensorflow as tf
 from tensorflow.keras.layers import Layer
 
-from merlin.models.tf.blocks.core.base import Block
-from merlin.models.tf.blocks.core.transformations import (
+from merlin.models.tf.blocks.retrieval.base import ItemRetrievalScorer
+from merlin.models.tf.blocks.sampling.cross_batch import PopularityBasedSampler
+from merlin.models.tf.core.base import Block
+from merlin.models.tf.core.transformations import (
     ItemsPredictionWeightTying,
     L2Norm,
     LabelToOneHot,
     LogitsTemperatureScaler,
     PopularityLogitsCorrection,
 )
-from merlin.models.tf.blocks.retrieval.base import ItemRetrievalScorer
-from merlin.models.tf.blocks.sampling.cross_batch import PopularityBasedSampler
 from merlin.models.tf.prediction_tasks.classification import (
     CategFeaturePrediction,
     MultiClassClassificationTask,
