+++ merlin/models/tf/prediction_tasks/retrieval.py
@@ -18,11 +18,11 @@ from typing import Optional, Sequence
 import tensorflow as tf
 from tensorflow.keras.layers import Layer
 
-from merlin.models.tf.blocks.core.base import Block
-from merlin.models.tf.blocks.core.transformations import LogitsTemperatureScaler
 from merlin.models.tf.blocks.retrieval.base import ItemRetrievalScorer
 from merlin.models.tf.blocks.sampling.base import ItemSampler
 from merlin.models.tf.blocks.sampling.in_batch import InBatchSampler
+from merlin.models.tf.core.base import Block
+from merlin.models.tf.core.transformations import LogitsTemperatureScaler
 from merlin.models.tf.metrics.topk import TopKMetricsAggregator
 from merlin.models.tf.prediction_tasks.classification import MultiClassClassificationTask
 from merlin.models.utils import schema_utils
